requirements.txt
requests==2.28.2
django==4.1.5
markdown==3.4.1
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
tqdm==4.64.1
